requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.26.0
bitsandbytes>=0.42.0

# VLM specific
qwen-vl-utils>=0.0.1
Pillow>=10.0.0

# Audio processing
faster-whisper>=0.10.0

# GraphRAG
graphrag>=0.1.0  # Microsoft's GraphRAG
llama-index>=0.9.0
faiss-cpu>=1.7.4  # or faiss-gpu
sentence-transformers>=2.2.0

# PDF processing
PyMuPDF>=1.23.0

# Web interface
gradio>=4.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0